requirements.txt
Flask
gunicorn
python-dotenv
flask-login
hashids